handling the non existing condition of the input video

+++ src/video_to_frames.py
@@ -17,36 +17,40 @@ import os
 
 def get_frames(video_input_path, frame_output_path):
 
-    # Playing the input video from file
-    video_capture = cv2.VideoCapture(video_input_path)
+    if (os.path.isfile(video_input_path)):
 
-    try:
-        if not os.path.exists(frame_output_path):
-            os.makedirs(frame_output_path)
-    except OSError:
-        print('Error: Creating directory of data')
+        # Playing the input video from file
+        video_capture = cv2.VideoCapture(video_input_path)
 
-    # Capture the very first frame
-    return_status, frame = video_capture.read()
+        try:
+            if not os.path.exists(frame_output_path):
+                os.makedirs(frame_output_path)
+        except OSError:
+            print('Error: Creating directory of data')
 
-    current_frame = 0
-
-    while(return_status):
-
-        # Saving the current frame's image as a jpg file
-        frame_location = frame_output_path+"frame" + str(current_frame) + ".jpg"
-        print ("Creating..." + frame_location)
-        cv2.imwrite(frame_location, frame)
-
-        # Capture frame-by-frame
+        # Capture the very first frame
         return_status, frame = video_capture.read()
 
-        # Increasing the current frame value for the next frame
-        current_frame += 1
+        current_frame = 0
 
-    # Release the capture
-    video_capture.release()
-    cv2.destroyAllWindows()
+        while(return_status):
+
+            # Saving the current frame's image as a jpg file
+            frame_location = frame_output_path+"frame" + str(current_frame) + ".jpg"
+            print ("Creating..." + frame_location)
+            cv2.imwrite(frame_location, frame)
+
+            # Capture frame-by-frame
+            return_status, frame = video_capture.read()
+
+            # Increasing the current frame value for the next frame
+            current_frame += 1
+
+        # Release the capture
+        video_capture.release()
+        cv2.destroyAllWindows()
+    else:
+        print("Invalid input video to capture. Location or the video not exist.")
 
 def run():
     video_input_path = "./data/input_video/input_video.mp4"
